fix nearby

+++ dto/response/health_facilities_response.go
@@ -10,6 +10,6 @@ type HealthResponse struct {
 	PhoneNum string
 	Name     string
 	Image    *string
-	Ranges   int
+	Ranges   float64
 	Address  model.Addresses
 }
